confirmar con contraseña

+++ src/views/Adm/Reusuarios.vue
@@ -1,6 +1,66 @@
 <template>
   <div>
     <div class="ma">
+
+      <!-- Modal -->
+      <div
+        class="modal"
+        id="Eliminar"
+        data-bs-backdrop="static"
+        data-bs-keyboard="false"
+        tabindex="-1"
+        aria-labelledby="staticBackdropLabel"
+        aria-hidden="true"
+      >
+        <div class="modal-dialog">
+          <div class="modal-content">
+          <!--Muestra el trailer con la url obtenida-->
+            <div class="row">
+              <h3>Para eliminar al usuario</h3>
+              <h3>{{nombre}} {{apellido}}</h3>
+              <h3>Debe colocar su contraseña</h3>
+              <h3>como Administrador</h3>
+              <div class="col-xs-12 col-sm-12 col-md-12">
+                <div class="form-floating">
+                  <input
+                    v-model="vacontraseña"
+                    type="text"
+                    name="vacontraseña"
+                    class="form-control"
+                    id="floatingInput"
+                    placeholder="Contraseña Actual"
+                    autocomplete="off"
+                  />
+                  <label for="floatingInput">Contraseña Actual</label>
+                </div>
+              </div>
+              <div class="alert alert-secondary" role="alert">
+                <label>{{ result }}</label>
+              </div>
+              <div class="col-xs-12 col-sm-12 col-md-12">
+                <a
+                  type="button"
+                  class="btnAdm0"
+                  data-bs-dismiss="modal"
+                  v-on:click="limpiar()"
+                >
+                  <i class="bx bxs-exit"></i>
+                </a>
+                <a
+                  v-on:click="eliminar(id, nombre)"
+                  type="submit"
+                  s
+                  class="btnAdm1"
+                >
+                  <i class="bx bx-trash-alt"></i>
+                </a>
+              </div> 
+            </div>
+          </div>
+        </div>
+      </div>
+      <!--fin Ventana Modal-->
+
       <!-- Modal -->
       <div
         class="modal fade"
@@ -13,135 +73,147 @@
       >
         <div class="modal-dialog modal-xl">
           <div class="modal-content">
-            <div class="modal-body">
-              <div class="panel-body">
-                <div class="row">
-                  <div class="col-xs-6 col-sm-6 col-md-6">
-                    <div class="form-floating">
-                      <input
-                        v-model="nombre"
-                        type="text"
-                        name="nombre"
-                        class="form-control"
-                        id="floatingInput"
-                        placeholder="Nombre"
-                        autocomplete="off"
-                      />
-                      <label for="floatingInput">Nombre</label>
-                    </div>
-                  </div>
-                  <div class="col-xs-6 col-sm-6 col-md-6">
-                    <div class="form-floating">
-                      <input
-                        v-model="apellido"
-                        type="text"
-                        name="apellido"
-                        class="form-control"
-                        id="floatingInput"
-                        placeholder="Apellido"
-                        autocomplete="off"
-                      />
-                      <label for="floatingInput">Apellido</label>
-                    </div>
-                  </div>
-                </div>
 
-                <div class="row">
-                  <div class="col-xs-6 col-sm-6 col-md-6">
-                    <div class="form-floating">
-                      <input
-                        v-model="telefono"
-                        type="text"
-                        name="telefono"
-                        class="form-control"
-                        id="floatingInput"
-                        placeholder="Telefono de usuario  [ejm: +503 4444-5555]"
-                        autocomplete="off"
-                      />
-                      <label for="floatingInput">Telefono de Usuario [ejm: +503 4444-5555]</label>
-                    </div>
-                  </div>
-                  <div class="col-xs-6 col-sm-6 col-md-6" >
-                    <div class="form-floating" v-if="this.id == ''">
-                      <input
-                        v-model="contraseña"
-                        type="text"
-                        name="contraseña"
-                        class="form-control"
-                        id="floatingInput"
-                        placeholder="Contraseña de Usuario"
-                        autocomplete="off"
-                      />
-                      <label for="floatingInput">Contraseña de Usuario</label>
-                    </div>
-                  </div>
+            <div class="row">
+              <div class="col-xs-6 col-sm-6 col-md-6">
+                <div class="form-floating">
+                  <input
+                    v-model="nombre"
+                    type="text"
+                    name="nombre"
+                    class="form-control"
+                    id="floatingInput"
+                    placeholder="Nombre"
+                    autocomplete="off"
+                  />
+                  <label for="floatingInput">Nombre</label>
                 </div>
+              </div>
+              <div class="col-xs-6 col-sm-6 col-md-6">
+                <div class="form-floating">
+                  <input
+                    v-model="apellido"
+                    type="text"
+                    name="apellido"
+                    class="form-control"
+                    id="floatingInput"
+                    placeholder="Apellido"
+                    autocomplete="off"
+                  />
+                  <label for="floatingInput">Apellido</label>
+                </div>
+              </div>
+            </div>
 
-                <div class="row">
-                  <div class="col-xs-6 col-sm-6 col-md-6">
-                    <div class="form-floating">
-                      <input
-                        v-model="usuario"
-                        type="text"
-                        name="usuario"
-                        class="form-control"
-                        id="floatingInput"
-                        placeholder="Nombre de Usuario"
-                        autocomplete="off"
-                      />
-                      <label for="floatingInput">Nombre de Usuario</label>
-                    </div>
-                  </div>
-                  <div class="col-xs-6 col-sm-6 col-md-6">
-                    <label>Rol de Usuario</label>
-                    <div class="form-group">
-                      <select
-                        v-model="rol"
-                        class="form-select"
-                        aria-label="Default select example"
-                        name="select"
-                        onchange=""
-                      >
-                        <option value="2">Tiene Rol de Usuario</option>
-                        <option value="1">Tiene Rol de Administrador</option>
-                      </select>
-                    </div>
-                  </div>
+            <div class="row">
+              <div class="col-xs-6 col-sm-6 col-md-6">
+                <div class="form-floating"  v-if="this.id == ''">
+                  <input
+                    v-model="usuario"
+                    type="text"
+                    name="usuario"
+                    class="form-control"
+                    id="floatingInput"
+                    placeholder="Nombre de Usuario"
+                    autocomplete="off"
+                  />
+                  <label for="floatingInput">Nombre de Usuario</label>
                 </div>
-                <div class="alert alert-secondary" role="alert">
-                  <label>{{ result }}</label>
+              </div>
+              <div class="col-xs-6 col-sm-6 col-md-6" >
+                <div class="form-floating" v-if="this.id == ''">
+                  <input
+                    v-model="contraseña"
+                    type="text"
+                    name="contraseña"
+                    class="form-control"
+                    id="floatingInput"
+                    placeholder="Contraseña de Usuario"
+                    autocomplete="off"
+                  />
+                  <label for="floatingInput">Contraseña de Usuario</label>
                 </div>
+              </div>
+            </div>
 
-                <div class="row">
-                  <div class="col-xs-12 col-sm-12 col-md-12">
-                    <a
-                      type="button"
-                      class="btnAdm0"
-                      data-bs-dismiss="modal"
-                      v-on:click="limpiar()"
-                    >
-                      <i class="bx bxs-exit"></i>
-                    </a>
-                    <a
-                      v-if="this.id == ''"
-                      v-on:click="setUsuarioApi()"
-                      type="submit"
-                      s
-                      class="btnAdm1"
-                    >
-                      <i class="bx bxs-save"></i>
-                    </a>
-                    <a
-                      v-if="this.id != ''"
-                      v-on:click="putUsuarioApi()"
-                      type="submit"
-                      s
-                      class="btnAdm1"
-                    >
-                      <i class="bx bxs-edit-alt"></i>
-                    </a>
-                  </div>
+            <div class="row">
+              <div class="col-xs-6 col-sm-6 col-md-6">
+                <div class="form-floating">
+                  <input
+                    v-model="telefono"
+                    type="text"
+                    name="telefono"
+                    class="form-control"
+                    id="floatingInput"
+                    placeholder="Telefono de usuario  [ejm: +503 4444-5555]"
+                    autocomplete="off"
+                  />
+                  <label for="floatingInput">Telefono de Usuario [ejm: +503 4444-5555]</label>
                 </div>
+              </div>
+              <div class="col-xs-6 col-sm-6 col-md-6">
+                <label>Rol de Usuario</label>
+                <div class="form-group">
+                  <select
+                    v-model="rol"
+                    class="form-select"
+                    aria-label="Default select example"
+                    name="select"
+                    onchange=""
+                  >
+                    <option value="2">Tiene Rol de Usuario</option>
+                    <option value="1">Tiene Rol de Administrador</option>
+                  </select>
+                </div>
+              </div>
+            <h1>Para Confirmar el cambio tiene que ingresar</h1>
+            <div class="col-xs-12 col-sm-12 col-md-12">
+              <div class="form-floating">
+                <input
+                  v-model="vacontraseña"
+                  type="text"
+                  name="vacontraseña"
+                  class="form-control"
+                  id="floatingInput"
+                  placeholder="Contraseña de Usuario Actual"
+                  autocomplete="off"
+                />
+                <label for="floatingInput">Contraseña de Usuario Actual</label>
+              </div>
+            </div>
+            </div>
+            <div class="alert alert-secondary" role="alert">
+              <label>{{ result }}</label>
+            </div>
+
+            <div class="row">
+              <div class="col-xs-12 col-sm-12 col-md-12">
+                <a
+                  type="button"
+                  class="btnAdm0"
+                  data-bs-dismiss="modal"
+                  v-on:click="limpiar()"
+                >
+                  <i class="bx bxs-exit"></i>
+                </a>
+                <a
+                  v-if="this.id == ''"
+                  v-on:click="setUsuarioApi()"
+                  type="submit"
+                  s
+                  class="btnAdm1"
+                >
+                  <i class="bx bxs-save"></i>
+                </a>
+                <a v-if="this.id != ''"
+                  v-on:click="contra()"
+                  type="submit"
+                  s
+                  class="btnAdm1"
+                >
+                  <i class="bx bxs-save"></i>
+                </a>
+                
               </div>
             </div>
           </div>
@@ -179,13 +251,16 @@
             <th>
               <a
                 class="btnAdm0"
-                v-on:click="eliminar(item.id, item.loginNameUser)"
+                type="button"
+                data-bs-toggle="modal"
+                data-bs-target="#Eliminar"
+                v-on:click="selec(item.id)"
               >
                 <i class="bx bx-trash-alt"></i>
               </a>
 
               <a
-                v-on:click="editar(item.id)"
+                v-on:click="selec(item.id)"
                 type="button"
                 class="btnAdm1"
                 data-bs-toggle="modal"
@@ -219,29 +294,54 @@ const Usuario = (()=>{
 
   //funcion
   const usuario = function(result, usu){
+
       // verifica que los campos nos esten vacios
       if (usu[0].nombre == "" || usu[0].apellido == "" || usu[0].telefono == "" ||
-      usu[0].contra == "" || usu.usuario == "" || usu.rol == "") {
+      usu[0].usuario == "" || usu[0].rol == "" || usu[0].vacontra == "") {
         
         // asigna respuesta
         result = "Verifique que los campos no esten vacios";
       }
 
-      else if(!telephoneCheck(usu[0].telefono)){
+      else if (!telephoneCheck(usu[0].telefono)){
         // asigna respuesta
         result = "Verifique que el telefono sea valido";
       }
+
       // de lo contrario
       else{
         // asigna respuesta
-        result = "Guardo con Exito"
+        result = "Correcto"
       }
     // retorna el result
     return result;
     
   }
+
+  const eliminar = function(result, pass, Adpass){
+
+    // verifica que los campos nos esten vacios
+    if (pass == "" || Adpass == "") {
+      // asigna respuesta
+      result = "Verifique que los campos no esten vacios";
+    }
+
+    // verifica que los campos nos esten vacios
+    if (pass != Adpass ) {
+      // asigna respuesta
+      result = "Verifique que los campos no esten vacios";
+    }
+
+    else{
+      result = "Correcto"
+    }
+
+    return result;
+
+  }
+
   // hace publica a la funcion
-  return{usuario}
+  return{usuario,eliminar}
 
 })();
 
@@ -258,6 +358,9 @@ export default {
       // variable de informacion
       result: "Recuerde llenar todos los campos",
 
+      //al macena la contraseña de la cookie
+      pass:  this.$cookies.get("pass"),
+
       // variables para almacenamiento de datos
       id: "",
       nombre: "",
@@ -266,10 +369,26 @@ export default {
       contraseña: "",
       usuario: "",
       rol: "",
+      vacontraseña: "",
     };
   },
   //METODOS A UTILIZAR
   methods: {
+    validar(){
+      // se almacena la informacion al arreglo
+      this.Vusu = [{
+                nombre:  this.nombre,
+                apellido:  this.apellido,
+                telefono:  this.telefono,
+                usuario:  this.usuario,
+                rol:  this.rol,
+                vacontra: this.vacontraseña
+              }]
+      //se alamacena lo que retorne de la funcion
+      //a la funcion se le envia la variable de informacion y el arreglo Vusu
+      this.result = Usuario.usuario(this.result, this.Vusu);
+    },
+
     tabla() {
       datatable;
 
@@ -310,21 +429,10 @@ export default {
     },
 
     setUsuarioApi() {
-      // se almacena la informacion al arreglo
-      this.Vusu = [{
-                nombre:  this.nombre,
-                apellido:  this.apellido,
-                telefono:  this.telefono,
-                contra:  this.contraseña,
-                usuario:  this.usuario,
-                rol:  this.rol,
-              }]
-      //se alamacena lo que retorne de la funcion
-      //a la funcion se le envia la variable de informacion y el arreglo Vusu
-      this.result = Usuario.usuario(this.result, this.Vusu);
+      this.validar();
 
       // se verifica el resutado obtenido
-      if (this.result == "Guardo con Exito") {
+      if (this.result == "Correcto") {
 
         axios
           // se indica la API a utilizar
@@ -342,11 +450,12 @@ export default {
             this.limpiar();
             //recarga el metodo
             this.getUsuariosApi();
+            this.result = "Guardado con Exito"
           });
       }  
     },
 
-    editar(id) {
+    selec(id) {
       this.getUsuariosApi();
 
       this.listaUsuarios.forEach((user) => {
@@ -356,21 +465,40 @@ export default {
             (this.apellido = user.lastNameUser),
             (this.telefono = user.phoneUser),
             (this.usuario = user.loginNameUser),
-            (this.contraseña = user.loginPasswordUser),
-            (this.rol = user.idRolUser),
-            console.log(user.firstNameUser);
+            (this.rol = user.idRolUser);
         }
       });
     },
 
+    contra(){
+      this.validar();
+      // se verifica el resutado obtenido
+      if (this.result == "Correcto") {
+        axios
+          .put("http://127.0.0.1:8000/api/users-change-password", {
+            loginNameUser: this.usuario,
+            oldPassword: this.vacontraseña,
+            newPassword: this.vacontraseña,
+          })
+          .then(() => {
+            this.result = "Editado con Exito"
+            this.putUsuarioApi()
+          }).catch(()=>{
+            this.result = "Contraseña incorrecta"
+          })
+      }
+    },
+
     putUsuarioApi() {
+    console.log(this.nombre);
+      
       axios
         .put("http://127.0.0.1:8000/api/users/" + this.id, {
           firstNameUser: this.nombre,
           lastNameUser: this.apellido,
           phoneUser: this.telefono,
           loginNameUser: this.usuario,
-          loginPasswordUser: this.contraseña,
+          loginPasswordUser: this.vacontraseña,
           idRolUser: this.rol,
         })
         .then(() => {
@@ -380,18 +508,23 @@ export default {
         })
         .catch(function (error) {
           console.log(error);
-          if (error) {
-            return (this.result = "REVISE QUE NO HAYA CAMPOS VACIOS");
-          }
+          return (this.result = "REVISE QUE NO HAYA CAMPOS VACIOS");
         });
+      
     },
 
     eliminar(id, NameUser) {
-      alert("ha eliminado al Usuario: " + NameUser);
 
-      axios.delete("http://127.0.0.1:8000/api/users/" + id).then(() => {
-        this.getUsuariosApi();
-      });
+      this.result = Usuario.eliminar(this.result, this.pass, this.vacontraseña);
+
+      if (this.result == "Correcto") {
+
+        alert("ha eliminado al Usuario: " + NameUser);
+
+        axios.delete("http://127.0.0.1:8000/api/users/" + id).then(() => {
+          this.getUsuariosApi();
+        });
+      }
     },
 
     limpiar() {
@@ -402,7 +535,9 @@ export default {
       this.contraseña = "";
       this.usuario = "";
       this.rol = "";
+      this.vacontraseña = "";
     },
+
   },
   //METODO INICIADO
   mounted() {
@@ -412,4 +547,8 @@ export default {
 };
 </script>
 
-<style lang="css"></style>
+<style lang="css">
+h1, h3{
+  color: black;
+}
+</style>
